Code refactor to make more readable
