+++ cmake/platform_win.cmake
@@ -14,12 +14,23 @@ function(set_source_files)
         MSVC/vasprintf-msvc.c
         MSVC/asprintf.c
         Fleece/Support/Backtrace+signals-win32.cc
+        Fleece/Support/Backtrace+capture-win32.cc
         PARENT_SCOPE
     )
 endfunction()
 
 function(set_base_platform_files)
-    # No-op
+    set(oneValueArgs RESULT)
+    cmake_parse_arguments(WIN_SSS "" "${oneValueArgs}" "" ${ARGN})
+    if(NOT DEFINED WIN_SSS_RESULT)
+        message(FATAL_ERROR "set_source_files_base needs to be called with RESULT")
+    endif()
+
+    set(
+        ${WIN_SSS_RESULT}
+        Fleece/Support/Backtrace+capture-win32.cc
+        PARENT_SCOPE
+    )
 endfunction()
 
 function(set_test_source_files)
